refactor(keystone): DOMA-13081 update ExternalContent field utilities and restructure imports
- Refactored ExternalContent field implementation by moving utility functions to a new module for better organization.
- Updated import paths for createExternalDataField and isFileMeta to reflect the new structure.
- Removed obsolete utils.js file to streamline the codebase.
